Add kernel testing support.
This requires a kernel built with CONFIG_GCOV_KERNEL=y and with debugfs mounted
on /sys/kernel/debug. You have to rename runtest.conf.example to runtest.conf
and edit it to specify your kernel build directory: coverage info for the kernel
then appears in the coverage/ subdirectory of the kernel tree, just as userspace
coverage info appears in the coverage/ subdirectory of the build-* tree(s).

Also quiet down extraneous output, so all that appears on stdout is test
coverage percentages, not progress info.