Fix expected-results corruption.
The expected-results postprocessor in the testsuite runner was corrupting
expected results that included non-columnar data after columnar output (e.g.
aggregate result printouts). Fixed by hunting for a colon in the 40th place
after every run of blank lines, and assuming that any line that does not include
a colon in that position is not in the columnar this-fired output format.