Compensate for desynchronization of stdout and stderr.
dtrace helpfully emits at least one line on stderr whenever it runs, even if the
run was successful (the 'matches N probes' line).  Output to stdout and stderr is
desynchronized (even writing to a file opened with O_APPEND does not help as one
is buffered in userspace and the other is not).  So, to ensure consistency of
expected results, we now segregate all output on stderr onto the end of the .r
file, after a line reading only "-- @@stderr --".

--capture-expected captures results in the expected format.  All existing
expected results have been respun accordingly.