Fix test/intro/rwtime.r.p.
This results postprocessor had a huge C copyright comment in it, though its
absence of a hashbang header meant that it was being treated as a bash script by
default. 35 bytes is way below any limit for copyrightability you care to set:
the header is unnecessary.

Also, it was trying to use $1 to reference the text matched by the address
selector in a sed script. sed is not awk and ${number} has no meaning in sed.
Fixed to use s/// instead.