Quiet testing.
With the number of new XFAILs, we are in danger of losing sight of actual
failures.  This change introduces a new --quiet mode to runtest.sh which
suppresses display of all XFAILs and PASSes, leaving only FAILs and XPASSes
(things which passed but which we expected to fail).  'make check' now runs
quiet mode by default, so all output from 'make check' indicates an actual
problem.  (Only actual FAILs trigger an exit with nonzero exitcode, because
XPASSes indicate a problem that is probably in the *testsuite*, not in
dtrace itself.)