Allow programmatic skipping of tests: skip some if loadavg is too high.
Until now, .x scripts have been an unused feature in the testsuite engine
permitting tests to examine their environment and XFAIL their corresponding test
by returning nonzero if they decide the test cannot pass in this environment.

We now generalize this: the scripts can now return 0 (sucess expected),
1 (XFAIL) or 2 (silently skip).  We use this to skip three tests which allocate
a lot of memory when the load average is high enough that it is likely that
processes are already blocked on I/O and thus the likely swap activity cannot
complete before the (already generous) test timeout.