Shared library support.
This is quite tricky.  We have multiple problems:

 - In-build-tree, we have multiple libraries: out-of-tree, we want only
   one (libdtrace.a/.so).

 - Those libraries have very dirty symbol namespaces.  We must apply a symbol
   version file to hide most of them.

 - Many of those libraries are only partially used: we don't want to include
   symbols in e.g. the libproc or libctf directories inside libdtrace which are
   not used by libdtrace itself.

 - Plus, of course, we want to hide as much as possible of the shared-
   library-building mess from the build system users.

The solution is to partition the concept of a 'library' into two:

 - a 'build library' is a .a file generated from some source files, which cannot
   be linked against (the build system mangles its name lightly to make it a bit
   harder to do by mistake).  Its files are all compiled with -fPIC.  This is
   done with the BUILDLIBS variable, whose use is just like the old LIBS
   variable (it's the same variable, renamed).

 - Linkable libraries are .a or .so files generated by combining one or more
   build libraries.  They can be linked against via _LIBS as usual.  They are
   created via the LIBS or SHLIBS variables, and have a new $name_LIBSOURCES
   variable specifying which build libraries go to make them up.

Shared libraries in particular have additional (optional) $name_VERSION,
$name_SONAME and $name_VERSCRIPT variables which can be used to specify the
library's full version, soname, and version script.  There's also a
$name_SECONDARY variable which names those build libraries whose symbols should
only be brought into the shared library if they are needed by part of the shared
library itself: it's a way of asserting 'these things are not separately
installed and contain no public symbols: leave unused stuff out'.

We can then make libproc a build library (as it is used only by
libdtrace). libctf and libport are used lightly by cmd/dtrace.c, so they have to
be declared as both build libraries and linkable libraries by putting them in
both BUILDLIBS and LIBS and setting their LIBSOURCES variables, e.g.

libctf_LIBSOURCES := libctf

('the libctf linkable library consists of all the symbols in the libctf build
library').

libdtrace itself gets declared to be a shared library only, combining all of the
libdtrace build library, libproc, libctf and libport, but with the latter three
as SECONDARYs so that unused code in them does not land in libdtrace.  We can
then install it from the libdtrace install target, and all is well.  (For now,
we call it libdtrace.so.0, since its API and ABI are not stable.)

We also need to update runtest.sh a bit to set LD_LIBRARY_PATH when testing a
non-installed dtrace.