Partition the testsuite for easier stress-testing.
This divides the tests into three sets:

 - unit tests under test/unittest, mostly already present but including the old
   test/dif and test/regression directories.

 - demo tests under test/demo, the old DTrace demo-derived tests, mostly relying
   on timeouts for termination.

 - stress tests under test/stress.  These tests place the system under stress, so
   should not be run in parallel with tests that also place the system under
   stress (such as OLT).

A new option --testsuites={suite,suite...} is added to runtest to tell it which
testsuites to run (by default, all of them), and check-*-stress targets are
added to the makefile to run these tests.

Because the number of testing targets is getting a little high, the targets have
been re-expressed using target-specific variables, so that most of the checking
targets can now be written in two short lines, with minimal repetition.  Still,
if we end up with too many more checking targets we might want to start
machine-generating all these target names.