Avoid comparisons when stress-testing.
The expected-results comparison done by the dtrace testsuite is somewhat
precarious at best, given that the tests were designed only to produce a
consistent exit code and many of the results can vary if e.g. a timer tick is
missed due to load.

So add an option to skip comparing results, comparing exitcode only, as the old
Solaris testsuite engine did (though we also consider a coredump to be a sign of
a bug, which it did not).  When skipping comparison, also convert XPASSes into
PASSes, because the only sign of many XFAILs is a failure in expected results,
which will be missed if results comparison is turned off.

Use this option in the check-*-stress targets.