Silent-by-default build.
This is a lot neater than what is done in the Linux kernel: rather than adding a
whole new rule-definition system just to add '@' to the front of every command,
we append --silent to MAKEFLAGS to turn command echoing off wholesale, then
just define a callable make function to print pretty output out when needed.

The only rough edge is that it is quite ugly to define installation rules now:
you have to double the number of lines in them (once to print out the list of
what is going to be installed, the next time to install it). This is probably
fixable fairly easily should it get too annoying.