Allow addition of members to structures with a specified offset.
The ctf_add_member() function lets you add a member with a specified name and
type to the end of a structure -- but it always uses the natural alignment of
the machine to do so, which is not sufficient when re-encoding DWARF, as not
all structures need be laid out that way.  So we allow the user to explicitly
specify the offset of the new member (in bits) from the start of the structure
via the new API function ctf_add_member_offset().