Tiny specfile cleanups.
Pearly's not working here anymore and neither her nor I are the authors of
DTrace.  At best we could be regarded as maintainers.  Add Kris, adjust the
%description accordingly.

Change %files to claim all of /usr/lib/dtrace, not just every file that
happens to be in it right now.  Declare /usr/share/doc/dtrace-*/* to be
documentation.