Add a crude symbol-checking infrastructure, like Autoconf's AC_CHECK_LIB.
This lets you check for the presence of a symbol on the build system by calling

$(eval $(call check-symbol-rule,NAME,symbol,library))

which automatically does a link test for this symbol and emits a line into
$(objdir)/config.h #defining HAVE_$(NAME) to 1 if that symbol is present.

To force config.h to be built whenever needed, it is forcibly added to the
dependency list of every C file.  (Because it is often indirectly included,
we cannot reasonably ask users to add the dependency themselves.  It never
changes after initial creation, so this somewhat extreme solution is
acceptable.)