Adapt to the absence of elf_getshdrstrndx() and elf_getshdrnum().
These symbols were introduced in elfutils 0.142 as aliases for the deprecated
elf_getshstrndx() and elf_getshnum().  Unfortunately our snapshot build system
has an older elfutils than that, and we have repeatedly fallen victim to the
absence of these symbols on that system.

So use the new autoconf-like symbol-checking machinery to #define the new names
to the old ones where necessary, and use the new names everywhere, rather than
trying to use the old symbols and frequently forgetting a few, leading to
snapshot build failures later on.