Provide an implementation of strrstr().
This is like strstr() only finds the last match of the needle in the haystack,
rather than the first.

Implement in terms of strstr() on the grounds that this is probably optimized by
the libc.

I have seen Unix-like OSes that already implement strrstr() before (though Linux
does not), so we use the new symbol-checking machinery to provide it only if the
OS does not already implement it.