Do not generate coverage info for configure-test compilations.
After doing a 'make coverage=yes' run, a file '-.gcno' was appearing at the top
level of the dtrace source tree.  This is because CFLAGS contained '--coverage',
and the configure-test compilations done by Makeconfig were substituting in
CFLAGS for a compilation of source code coming from standard input.

We don't want coverage for that sort of thing, and the -.gcno file is messy, so
unconditionally pass --no-coverage to such compilations.