Prevent the lexer from looking up unknown types unnecessarily.
The first time we look up a type and don't find it is very expensive: every
module's CTF section must be loaded in off disk.  The lexer has several places
in which it unnecessarily checks to see if things that cannot be types are
types: this commit fixes two of them.

Firstly, the S2{RGX_PSPEC} pattern which matches probes and variable
declarations in the global scope looks up everything which does not contain a
colon to see if it might be a type (except when specified directly on the
command-line, where a 'pspec' option is flipped on by cmd/dtrace.c to prevent
this).  This is reasonable for things like 'int64_t', and if people specify
probes without giving the provider or module name they deserve a performance
hit... but 'BEGIN', 'END', and 'ERROR' cannot be specified with colon notation
but trigger a useless type lookup every time nonetheless.  So we explicitly
skip those three names now: we know they cannot be types.

Secondly, id_or_type().  The comment says "The code here is ordered carefully as
lookups are not cheap".  If by this they meant "ordered as expensively as
possible", perhaps this is true.  Almost the first thing id_or_type does is
checks to see if the lexeme it's looking at is a type.  This includes things
that cannot possibly be types, such as 'varname' in

  int64_t varname;

and NULL in

  inline long NULL = 0;

which we see in _dt_hardwire, the very first line of D ever executed.  Oops.
(In fact every single variable that ever gets initialized gets looked up as a
type first, even though that would always be a syntax error. It is doubly ironic
that id_or_type() actually contains a comment to the effect that that it is
pointless to check if things are types if that would be a syntax error -- right
after it does just that.)

Fixing this while not breaking the property that the first argument of
'offsetof' must be a type is quite tricky.  We also have to take care that
typedef names are considered identifiers when they are declared without pulling
them into the identifier hash (because after this point, we will consider them
type names).

In the process we fix a couple of other bugs, notably that the old id_or_type()
sometimes called input() then unput(), input()ting whitespace first in a loop:
but it forgot to unput() the whitespace again.  You might think this is
irrelevant, only #pragma and other such tokens depend on whitespace
(specifically linefeed) as delimiters.

This is enough to let us handle a lot of simple cases, including variable
declarations, without looking up unknown types and triggering mass module loads
and CTF reads.  Unfortunately 'struct task_struct' still triggers a type lookup
for 'task_struct', which does not exist (only 'struct task_struct' does).  This
will be fixed shortly (and a testcase added).

Finally, we fix a piece of flex-specific undefined behaviour: S2{RGX_PSPEC}
reads from yytext while calling unput(), which is permitted in POSIX but in flex
runs into the problem that unput() corrupts yytext.  The fix is simple: since we
already have a strdup()ped copy of yytext in any case, we can just work over
that instead.

The revised unput()ting changes some error messages a bit: those testcase
results are regenerated.