runtest: Support modules that are only unloaded, not loaded.
The testers verify that after the testsuite has run, no modules are loaded that
were loaded before the testsuite has started.  This test is fairly useless, but
as long as they do it, we can use it to verify properties that are actually
important -- to wit, that module dependencies are correct.

So we add a new keyword to test/modules, @@unload-only.  Modules marked with
this keyword will be unloaded (in reverse order), but never loaded. We mark the
dtrace_ctf and dtrace modules as @@unload-only, thus verifying that module
dependencies are correct and that all DTrace modules can be safely unloaded in
one fell swoop.