Avoid errors after a parser restart.
On error, we longjmp() out of the parser, and then -- if this is a .d file in
the library directory -- keep going to process further .d files.  Bison,
unlike AT&T lex, requires a yyrestart() in this situation.

We put this in yyinit() (so that it is done on all parser starts, not just
restarts) because it is much easier to just restart unconditionally than to
determine whether this is an error-induced restart and restart only then.