runtest: Add @@no-xfail.
This option is used to say that a test is expected to succeed, even though a
test.options above it may say that it is expected to fail.

This can be used to indicate that err.* tests are still expected to emit errors
even though the whole directory in which they reside is @@xfailed.

(Such tests in one directory, unittest/profile-n, are converted too.)