runtest: Revamp success/failure detection.
The success/failure detection code in runtest.sh was incredibly convoluted,
repetitious, and littered with bugs.  Notably, a result comparison failure was
considered so important that it overrode everything else, including reporting
and capturing of coredumps, and exits with nonzero exitcodes.  This is very
wrong: both of these conditions are more important and should be reported in
preference to a results comparison failure, since they are certain bugs while a
results comparison failure is only a possible bug.  This is complicated by the
need to include an expected-result diff in the .sum file whenever an .r file is
present, even if some other error was reported.

This refactoring fixes all these bugs, makes the code much easier to follow, and
does not negatively affect .sum or .log output in my testing.