Unknown-type-lookup reduction: struct/union dereferencing.
The next stage of unknown-type-lookup reduction is to teach the lexer that the
right-hand-side of -> and . operators is always an identifier, never a type
name.  We do this in much the same way that commit df5bac arranged for the word
after 'struct' / 'union' / 'enum' to always be a type name: we introduce a new
flag in the parser control block to track whether we are immediately after a
dereference of some sort, and test-and-clear it in id_or_type(), which is always
called for everything that looks even vaguely like an identifier or type name
and so is always going to be called at the right time.

We extend the existing test for unnecessary type-induced module loading to test
for this case too.