Omit test-coverage LDFLAGS where necessary.
Commit 5318a63 added --coverage to LDFLAGS when test coverage support is on, to
prevent shared-library builds from failing.  Unfortunately this meant that the
filtering-out of --coverage from CFLAGS that the triggers, the utils in
test/utils and the Makeconfig fragments were running was no longer effective.
For the former two cases this is not particularly bad, though inelegant: these
targets are separately linked, and --coverage in LDFLAGS without --coverage in
CFLAGS is harmless.  But for the Makeconfig fragments it meant that we ended up
with a spurious -.gcno file in the root of the tree again.

Makeconfig is easily fixed by extending the existing filtering-out of --coverage
to cover LDFLAGS as well.

The spurious (but harmless) --coverage in the other two cases is slightly
trickier to fix, requiring adding a *_NOLDFLAGS target-specific variable just
like the *_NOCFLAGS variable that is already used to filter --coverage out of
the CFLAGS for triggers and utils, using that variable in the functions in
Makefunctions that are used to link shared libraries and commands, and setting
it in test/triggers/Build and test/utils/Build.