Support GCC-preprocessor-style line directives.
GCC's preprocessor emits line directives in a format quite similar to the
Solaris preprocessor, but not identical.  The Solaris preprocessor has a single
integer following the filename to indicate whether we are entering or leaving an
inclusion.  The GCC preprocessor extended this format long ago to support a list
of integers, allowing extra flags to be specified indicating things like whether
this is a system header or whether it is implicitly extern "C".  DTrace doesn't
care about any of that -- it merely wants to track the header inclusion tree --
but it must know enough about it to be able to scan the list of integer flags
in case the nesting flags are at non-initial positions.