Fix non-loaded kernel module handling.
Non-loaded kernel modules are distinct entities from both loaded kernel modules
and user modules.  Like the former, their type information comes from CTF
sections in kernel modules, and can inherit from the shared dtrace_ctf module:
like the latter, symbol lookup must be done via ELF operations rather than using
the kernel symbol table.  Unlike both, symbol->address lookup cannot be expected
to work.

So implement this using a new module flag, DT_DM_KERN_UNLOADED, This gets set by
the call to dt_kern_module_find() in dt_module_load().  If a module exists in
the kernpath hash but has not already been marked as a kernel module, it cannot
have been found in the /proc/kallmodsyms scan and thus must not be
loaded.  (Conversely, if it's not found in the kernpath hash, it is a user
module, so the existing code that attempted to derive its types from dtrace_ctf,
as is done for builtin kernel modules, was incorrect.)

Loading of non-builtin modules is more like loading of a user module than
loading of a kernel module, sucking in all the usual ELF paraphernalia: if
we don't have kernel-provided address info we can get some of it at least
from ELF.  (Currently we don't use this for much, but in future it should be
possible to do variable->type lookups, which is currently only implemented
for loaded kernel modules.)

Add tests to ensure that type lookup from such a module works and variable
lookup does not.  (We use CAN as the 'unloaded module': it's part of OEL but it
is highly unlikely to run on anything but a car or an automotive testbed
system.)