runtest: log skips.
The @@skip option for skipped tests is too silent.  If skips are not at least
logged in runtest.log, nobody will ever remember to unskip them again.  (This
also means that we can log any message anyone has added to the @@skip, which
was previously thrown away and logged nowhere.)

Mark test/unittest/misc/tst.include.sh as @@skip: it needs porting, at the very
least, and is quite possibly doing nothing that makes sense on a Linux box at
all.  (And it takes so long to run that it often gets timed out, making @@xfail
not work to hide it since it then tends to show up as XPASS on slower systems.)