Unknown-type-lookup reduction: array dimensions and type compilation.
When an array dimension is seen, it can be of form

[53]

or

[int]

or even

[volatile const womble_t * const * restrict]

However, in an inline declaration we can see

[int bar]

or even

[volatile const womble_t * const * restrict baz]

in which that last bar or baz is an identifier, to be used in the
initializer (often in an xlate clause).

We can distinguish these cases by noting that in all but the identifier cases,
there may be no type present (as in '53') or there may be one type present, but
there cannot be more than one: foo[int * char] is not valid, and if we see
foo[int * bar] we know 'bar' must be an identifier.

So handle this by introducing a new flag in the parser control block which
indicates when we are inside array dimensions, and flipping it on and off
appropriately in the 'array' grammar rule.  Then, in the lexer's always-
problematic id_or_type() function, we can easily check to see if we are inside
array dimensions and if we have seen at least one type in this scope so far.  If
we have, we intern the identifier in the same temporary pcb_idents stash as is
used for variable declaration lexing.  (As with variable declarations, the
actual stashing of the type is done elsewhere: this stashing is purely to help
id_or_type() recognize that uses of this identifier in the initializer cannot be
types, something else it would otherwise get wrong.)

A semi-related change: when asking the kernel about its exported probes, the
types of the probes are compiled by calling dtrace_type_strcompile(), which
eventually turns into a call into the parser in pcb_context DT_CTX_DTYPE.  Since
we already know these types must be types from the parser context, return that
straight away rather than having the kernel force us to dig through oodles of
modules to determine that the types of D functions and translators are in fact
types.