Work on systems with no kernel source in /lib/modules/`uname -r`/build.
Commit c5153cee added a build dependency on the kernel headers, for substituting
constants in translators.  Unfortunately there is no guarantee that
/lib/modules/`uname -r`/build points to a DTrace-capable kernel, or any kernel
source at all: and because this is gcc-E-substituted, errors just lead to non-
substitution of values and failure at runtime.

So we introduce a new KERNELDIR variable and enforce that it points to something
that looks likely to be kernel source.  Setting this in the RPM specfile is more
interesting: we do not want to have to bump the userspace specfile whenever a
new kernel is released.  Instead, we introduce two new %kver_min and %kver_max
defines, which give version bounds on the kernel this userspace works with, and
use those to require that *some* appropriate devel package is installed: then we
use wildcard expansion to point KERNELDIR to any of those packages at build
time.