Name unnamed arguments in libdtrace public API header.  Export missing types.
The DTrace public API header file libdtrace/dtrace.h did not name any of its
arguments.  Since the source code to the implementation of these functions is not
available, and there is no other documentation, this is not ideal: many of the
arguments' purposes are clear from their types, but by no means all.

In particular, this hid a bug: dtrace_aggregate_print() takes a pointer to an
aggregate walk function, but generates the arguments to pass in itself, and uses
an internal type for that.  A value of NULL for the walk function means to use
the default, which works, but it is impossible to pass in a non-default value
since you cannot write one of these functions yourself.  Fixed by moving struct
dt_print_aggdata into dtrace.h (and renaming it dtrace_print_aggdata, since it
is -- and has in effect always been -- part of the public API.)