Add ps_prochandle.ptraced to track the ptrace state of processes.
As of commit 4fcb15b1, whether we PTRACE_DETACH from a process once it is set
running depends on the value of a command-line parameter to Psetrun().  This
means that we could not tell if further ptrace()s required a reattachment or
not, but this was not a problem since we never needed to do any further
ptrace()s.

As of a forthcoming commit this will no longer be true, so add a ptraced flag to
indicate whether we are PTRACE_ATTACHed to this process at present.