Massive cleanup of libproc/Psymtab.c ELF file acquisition.
Before now, the ELF file acquisition code in libproc's Psymtab.c was extremely
baroque: if the object file could not be opened, it entered a fabulous
wonderland of architecture- and operating-system-dependent code which used
rd_get_dyns() from the Solaris librtld_db API to extract the dynamic symbol
table from the in-core image and construct a fake ELF file using that data.

Even if it could be opened, another mass of operating-system-dependent madness
ensued: DTrace scanned for a DT_CHECKSUM section injected by the Solaris linker
so as to detect if the ELF file had been unlinked and recreated since the
traced process was started.

This was all later broken even more badly during porting by the false assumption
that Pbuild_file_symtab() was only called on the mapping corresponding to the
primary executable.  This is not so: it can be called for any executable
mapping.

Rather than all this nonsense, we can just use PTRACE_GETMAPFD to extract a file
descriptor to the file backing the mapping the process is *actually* using,
whether it has been deleted or not.  This is always the complete file, so it
already contains all ELF headers necessary: there is no need to fake anything.
The only complexity is that we may possibly need to PTRACE_ATTACH to the process
first, and PTRACE_DETACH afterwards, if we are not already tracing it.  (This is
done with the aid of the ps_prochandle.ptraced flag added in da0b649.)

This can handle everything Solaris DTrace could handle with the single exception
of code that creates an anonymous mapping and then builds an ELF file in it
dynamically and executes it.  This seems like a rather unlikely case: we'll
handle it if it ever turns up.  (It seems more likely that most such code will
just use an unlink()ed temporary file, which since it is a file-backed mapping
can be returned by PTRACE_GETMAPFD perfectly well.)